refactor: preparation for multiple zombie event finder processes
