+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/zombieevents/Generators.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.renku.eventlog.subscriptions.zombieevents
+
+import ch.datascience.graph.model.EventsGenerators.{compoundEventIds, eventStatuses}
+import ch.datascience.graph.model.GraphModelGenerators.projectPaths
+
+private object Generators {
+  lazy val zombieEvents = for {
+    eventId     <- compoundEventIds
+    projectPath <- projectPaths
+    eventStatus <- eventStatuses
+  } yield ZombieEvent(eventId, projectPath, eventStatus)
+}
