+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/zombieevents/ZombieEventEncoderSpec.scala
@@ -23,6 +23,7 @@ import ch.datascience.graph.model.EventsGenerators._
 import ch.datascience.graph.model.GraphModelGenerators.projectPaths
 import io.circe.literal._
 import io.renku.eventlog.subscriptions.EventEncoder
+import io.renku.eventlog.subscriptions.zombieevents.Generators.zombieEvents
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
 
@@ -31,7 +32,7 @@ class ZombieEventEncoderSpec extends AnyWordSpec with should.Matchers {
   "encodeEvent" should {
 
     "serialize ZombieEvent to Json" in {
-      val event = ZombieEvent(compoundEventIds.generateOne, projectPaths.generateOne, eventStatuses.generateOne)
+      val event = zombieEvents.generateOne
 
       encoder.encodeEvent(event) shouldBe json"""{
         "categoryName": "ZOMBIE_CHASING",
@@ -48,9 +49,7 @@ class ZombieEventEncoderSpec extends AnyWordSpec with should.Matchers {
   "encodePayload" should {
 
     "return None" in {
-      encoder.encodePayload(
-        ZombieEvent(compoundEventIds.generateOne, projectPaths.generateOne, eventStatuses.generateOne)
-      ) shouldBe None
+      encoder.encodePayload(zombieEvents.generateOne) shouldBe None
     }
   }
